Refactor data list.
